Move the helper functions into an underscore namespace
